refactor: move package for UserCommonTests
