+++ src/test/java/techcourse/myblog/controller/common/UserCommonTests.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.controller;
+package techcourse.myblog.controller.common;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
